1/resolve packet type error in class newsprofile.

+++ squirrel-client/src/main/java/cc/mashroom/squirrel/client/storage/model/chat/NewsProfile.java
@@ -39,7 +39,7 @@ public  class     NewsProfile
 	private  Timestamp   createTime;
 	@JsonProperty( value="PACKET_TYPE" )
 	@Column( name="PACKET_TYPE" )
-	private  Short    packetType;
+	private  Integer  packetType;
 	@JsonProperty( value="CONTACT_ID"  )
 	@Column( name="CONTACT_ID"  )
 	private  Long  contactId;
